Update README.md
It's a hackathon project which is meant to solve the genuine problem of movie lovers.  
We can never get a correct suggestion about movies and series from somewhere. So without depending on people or other websites, why not use our website "MOVIE DADDY" to find the best of the best recommendations of movies and series you love? 
                       
               Never get any biased or inappropriate suggestions about movies from people. Choose your own.  Our website uses machine learning algorithm to find you the best recommendations of movies and web-series you can ever get.